Add files via upload
This is an updated log file.
The new features apply to the logging site, where I added delete and reorder features and a personalized style to the HTML and JS files. I hope it's okay. The tests show that to reorder, you must have the words of the log fall below the one you want it beneath for it to work.